data_platform/shopify_etl/requirements.txt
google-cloud-firestore==2.11.1
google-cloud-bigquery==3.11.0
python-dateutil==2.8.2
requests==2.31.0
